+++ AeSys/EoMfOutputDockablePane.cpp
@@ -29,14 +29,14 @@ int EoMfOutputDockablePane::OnCreate(LPCREATESTRUCT createStruct) {
   EmptyRect.SetRectEmpty();
 
   if (!m_wndTabs.Create(CMFCTabCtrl::STYLE_FLAT, EmptyRect, this, 1, CMFCTabCtrl::LOCATION_BOTTOM)) {
-    ATLTRACE2(atlTraceGeneral, 0, L"Failed to create output tab window\n");
+    ATLTRACE2(static_cast<int>(atlTraceGeneral), 0, L"Failed to create output tab window\n");
     return -1;
   }
   const DWORD SharedStyles = WS_CHILD | WS_VISIBLE | WS_HSCROLL | WS_VSCROLL | LBS_NOINTEGRALHEIGHT;
 
   if (!m_OutputMessagesList.Create(SharedStyles, EmptyRect, &m_wndTabs, 2) ||
       !m_OutputReportsList.Create(SharedStyles, EmptyRect, &m_wndTabs, 4)) {
-    ATLTRACE2(atlTraceGeneral, 0, L"Failed to create output windows\n");
+    ATLTRACE2(static_cast<int>(atlTraceGeneral), 0, L"Failed to create output windows\n");
     return -1;
   }
   m_OutputMessagesList.SetFont(&m_Font);
